+++ src/views/account/helpers/accountDelegationsHelper.js
@@ -1,10 +1,12 @@
 const accountDelegationsHelper = {
   build({ delegations, validators }) {
-    return delegations.map((delegation) => ({
-      moniker: _getMoniker(validators, delegation.delegation),
-      operator: delegation.delegation.validator_address,
-      amount: parseFloat(delegation.balance.amount),
-    }));
+    return delegations
+      .filter((delegation) => parseFloat(delegation.balance.amount) > 0)
+      .map((delegation) => ({
+        moniker: _getMoniker(validators, delegation.delegation),
+        operator: delegation.delegation.validator_address,
+        amount: parseFloat(delegation.balance.amount),
+      }));
   },
 };
 
@@ -12,7 +14,7 @@ export default accountDelegationsHelper;
 
 const _getMoniker = (validators, delegation) => {
   const validator = validators.find(
-    (validator) => validator.operator_address === delegation.validator_address,
+    (validator) => validator.operator_address === delegation.validator_address
   );
   return validator ? validator.description.moniker : '-';
 };